gmsh wrapper for making a 2d triangular mesh

+++ api/gmshWrap.py
@@ -357,7 +357,7 @@ def gmsh2R2mesh(file_path='ask_to_open',save_path='default',return_mesh='no'):
             #the information here is returned as a mesh dictionary because it much easier to debug 
         
 #%% gmsh wrapper
-def tri_mesh(surf_x,surf_y,elec_x,elec_y,doi=50):
+def tri_mesh(surf_x,surf_y,elec_x,elec_y,doi=50,keep_files=False):
     """ generates a triangular mesh for r2. returns mesh.dat in the Executables directory 
     this function will only work if current working directory has path: Execuatbles/gmsh.exe"""
 #INPUT: 
@@ -387,7 +387,8 @@ def tri_mesh(surf_x,surf_y,elec_x,elec_y,doi=50):
     os.system('gmsh '+file_name+'.geo -2')#run gmsh 
     #convert into mesh.dat 
     mesh_dict=gmsh2R2mesh(file_path=file_name+'.msh',return_mesh='yes')
-    os.remove("temp.geo");os.remove("temp.msh")
+    if keep_files is False: 
+        os.remove("temp.geo");os.remove("temp.msh")
     #change back to orginal working directory
     os.chdir(cwd)
